Move all handlers into a single module and rely on reth builtin tx type detection
