+++ crates/full-node/sov-ethereum/src/lib.rs
@@ -1,10 +1,9 @@
 mod gas_price;
-#[cfg(feature = "local")]
-mod signer;
+mod handlers;
 
 use std::sync::Arc;
 
-use alloy_primitives::{Bytes, B256, U256};
+use alloy_primitives::B256;
 use jsonrpsee::types::{ErrorCode, ErrorObjectOwned};
 use jsonrpsee::RpcModule;
 pub use reth_rpc_eth_types::GasPriceOracleConfig;
@@ -14,12 +13,10 @@ pub use sov_eth_dev_signer::DevSigner;
 pub use sov_evm::EthereumAuthenticator;
 use sov_evm::{convert_to_transaction_signed, Evm, RlpEvmTransaction};
 use sov_modules_api::capabilities::HasKernel;
-use sov_modules_api::{ApiStateAccessor, RawTx, Spec};
+use sov_modules_api::{ApiStateAccessor, Spec};
 use sov_sequencer::Sequencer;
 
 use crate::gas_price::gas_oracle::GasPriceOracle;
-
-const ETH_RPC_ERROR: &str = "ETH_RPC_ERROR";
 
 #[derive(Clone)]
 pub struct EthRpcConfig {
@@ -63,58 +60,17 @@ where
     S::Address: FromVmAddress<EthereumAddress>,
     Seq::Rt: HasKernel<S> + EthereumAuthenticator<S> + Default + Send + Sync + 'static,
 {
-    rpc.register_async_method("eth_gasPrice", |_, ethereum, _| async move {
-        let price = {
-            let mut state = ethereum.api_state_accessor();
-
-            let suggested_tip = ethereum
-                .gas_price_oracle
-                .suggest_tip_cap(&mut state)
-                .await
-                .unwrap();
-
-            let evm = Evm::<S>::default();
-            let base_fee = U256::from(
-                evm.get_block_by_number(None, None, &mut state)
-                    .unwrap()
-                    .unwrap()
-                    .header
-                    .base_fee_per_gas
-                    .unwrap_or_default(),
-            );
-
-            suggested_tip + base_fee
-        };
-
-        Ok::<U256, ErrorObjectOwned>(price)
-    })?;
-
-    rpc.register_async_method(
-        "eth_sendRawTransaction",
-        |parameters, ethereum, _| async move {
-            let data: Bytes = parameters.one().unwrap();
-
-            let raw_evm_tx = RlpEvmTransaction { rlp: data.to_vec() };
-
-            let (tx_hash, raw_message) = ethereum
-                .make_raw_tx(raw_evm_tx)
-                .map_err(|e| to_jsonrpsee_error_object(e, ETH_RPC_ERROR))?;
-
-            let tx = Seq::Rt::encode_with_ethereum_auth(RawTx::new(raw_message));
-
-            ethereum.sequencer.accept_tx(tx).await.map_err(|e| {
-                to_jsonrpsee_error_object(
-                    format!("{} - '{}' ({:?})", e.status, e.message, e.details),
-                    ETH_RPC_ERROR,
-                )
-            })?;
-
-            Ok::<_, ErrorObjectOwned>(tx_hash)
-        },
-    )?;
+    rpc.register_async_method("eth_gasPrice", handlers::eth_gas_price)?;
+    rpc.register_async_method("eth_sendRawTransaction", handlers::eth_send_raw_transaction)?;
 
     #[cfg(feature = "local")]
-    signer::register_signer_rpc_methods::<S, Seq>(rpc)?;
+    {
+        rpc.register_async_method("eth_accounts", handlers::signer::eth_accounts)?;
+        rpc.register_async_method(
+            "eth_sendTransaction",
+            handlers::signer::eth_send_transaction,
+        )?;
+    }
 
     Ok(())
 }
